requirements.txt
fastapi
uvicorn
openai
langchain
chromadb
pydantic
python-dotenv
